requirements.txt
anyio==3.6.2
blinker==1.4
click==8.1.3
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dlib==19.24.99
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.95.1
h11==0.14.0
httplib2==0.20.2
httptools==0.5.0
idna==3.4
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==1.24.2
oauthlib==3.2.0
Pillow==9.5.0
pydantic==1.10.7
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.3.0+ubuntu2.1
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
systemd-python==234
typing_extensions==4.5.0
ubuntu-advantage-tools==27.12
ufw==0.36.1
unattended-upgrades==0.1
uvicorn==0.21.1
uvloop==0.17.0
wadllib==1.3.6
watchfiles==0.19.0
websockets==11.0.1
zipp==1.0.0
